Made AbstractReplyMessageReceiver non-abstract (and renamed to ReplyMessageReceiver)
Removed empty subclasses and inserted ReplyMessageReceiver into the configuration parser, with one exception:
WebServiceReplyMessageReceiver remains since it is used as a type in TestNGCitrusTestBuilder for the Java DSL so that
it can create a different action for SOAP Messages. Maybe there is a better solution so that this Receiver can vanish, too ?

Also, quite some Unit-Tests (like JmxReplyMessageReceiverTest) could be removed, since they are all duplicates now.
However, this is not included in this commit.