Fix synchronization issues in context variables.
Variables were stored in LinkedHashMap, which is not thread safe. If a
test uses parallel actions, and one thread stores a variable which leads
to an increase of size of the LinkedHashMap, while another thread reads
a variable, this leads to a (sporadial) "Unknown variable" exception.
This can be prevented by using the threadsafe ConcurrentHashMap.