Updated HttpMessageContentBuilder to account for variable substitution in headers.
The existing implementation did not account for variable substitution in header keys.
In Citrus < 2.7.2 variable substitution for maps only worked on entries, not keys, so after substituting values in headers, existing entries would be overwritten.
With variable substitution for keys in place this no longer works and previous headers have to be cleared, otherwise there'd be one header containing ${variables} and a separate one with actual values.

This in fact can cause HTTP 400 - Bad request errors.