refactor(citrus-3.0): Move citrus-core unit tests to individual modules
Continue to modularize citrus-core by moving all unit tests to the modules they belong to. This requires us to introduce some amount of redundancy regarding the unit test base class. In order to avoid circular dependencies in modules we need to add unit test base classes in multiple modules.

In addition to that continue to switch to using the reference resolver instead of autowiring dependencies and using the Spring application context in components. This will allow us to operate in a non Spring mode.