refactor(citrus-3.0): Resolve strong dependencies to validation modules in citrus-core-spring
Core Spring module has had strong dependencies to validation modules (e.g. citrus-validation-xml, citrus-validation-hamcrest) in order to add message validator implementations and validation matchers to the application context configuration. This PR resolves the dependency by adding a resource path lookup mechanism to automatically add message validators and matchers when these are present on the classpath. This way the user can decide which validation modules to add in a project.

The refactoring required to use reference resolver instead of Spring application context in several components (e.g. soap fault message validators). The decoupling of the Spring application context in favor of using resource lookup or reference resolver is a secondary goal that needs to be continued in further refactorings.

Because of the much cleaner module dependency graph a bug in citrus-mail dependencies has been discovered. The javax.mail api and implementation versions were incompatible and only worked because of a transitive dependency to an older javax.mail-api version in citrus-validation-json module. Fixed this dependency issue by explicitly adding javax.mail dependency in compliant version and excluding the transitive dependency.