refactor(citrus-3.0): Introduce new Java DSL fluent API
New Java DSL fluent builder API should combine the best of test designer and test runner. The fluent API should work without lambda expressions like designer API. At the same time we want to have immediate action execution as test runner is doing. Latter is because we want to be able to use custom code in actions and avoid a separation of design time and runtime. Also we would like to be able to debug the test case and have exception stack traces pointing to the line of failure.

This is the first PR starting to implement the new fluent API for Java DSL.