refactor(citrus-3.0): Fix wait for async container for finish
When the test case is finished it waits for asynchronous actions to finish properly. This mechanism was broken for async test container resulting in failed tests because the async action did not report when it is done leading to a timeout exception while waiting for the test case to finish properly. This was due to a racing condition in executing nested test actions within the async container while reporting the done state to the test case. In some cases the list of executed actions in the async container still was changing while the async container tries to report if all nested actions are done. This resulted in a ConcurrentModificationException while reporting the done state. The done state polling mechanism did break because of that exception and never recovered from that state so the test failed because it did not finish properly.

Now we catch such a concurrent modification exception and any other exception that may be raised while polling for the done state of test actions. In case the exception was of temporary nature because of racing conditions the next poll may be successful. If the exception does not recover the test fails because the done state could not be determined. This fixes test cases failing to wait for done state when using async containers or wait actions.