refactor(citrus-3.0): Fix false positives when validating messages during ReceiveMessageAction
When receiving a message a validation is performed according to provided control message content. Some validations (e.g. Json, Xml, Xpath, message header) require a validation context to be present. The receive message action builder was not adding the message validation context properly for message objects. This means when user gives an expected message object the builder had problems to determine the right validation contexts to add. This is the case for modules like JDBC that act on message object validation a lot. This resulted in missing validation contexts and message validation was skipped resulting in false positives.

Improved the receive message action builder to reconcile the validation context list before the final build step. This ensures the proper setting of message validation contexts in the receive message action.