refactor(citrus-3.0): Extract CitrusEndpoints from citrus-java-dsl module
CitrusEndpoints combine all available endpoint builder DSL methods in a single class. Users can use this as a catalog of available endpoints when creating endpoints in Java DSL test cases. This PR extracts the endpoint catalog into a separate module. This way the user can add it to the project if needed. The "vintage" Java DSL module will add this by default.