refactor(citrus-3.0): Fix unstable SyncJmsTopicJavaIT integration tests
The issue is a runtime race condition on the two parallel receive operations on the same synchronous topic endpoint. Each receiver stores its temporary reply topic with different message ids to the message correlation manager. This can lead to inconsistent state in correlation manager and the reply topic can not be found. The fact to receive the very same message twice in parallel form the very same synchronous topic is very specific to the sample test case and is not a common scenario someone will have in production. Fixed the race condition by adding a sleep to one of the receive operations.