refactor(citrus-3.0): Improve message validator lookup in empty messages
When empty message payload is provided the message validator registry had problems to pick the right validator and often raised an error that no proper validator was found. This PR takes care on empty messages that are validated automatically when no proper validator was found. This makes the mechanism more robust and there is no need to include a message validator only for empty message purpose.