refactor(citrus-3.0): Move to Jakarta EE
Moves to using Jakarta EE dependencies instead of `javax` or `com.sun` ones. Because of many dependencies we pull in that also use the `javax` or `com.sun` dependencies this also tries to conform those to the Jakarta EE ones by excluding the `javax` or `com.sun` ones from the dependency tree.

Fixes #712