chore(build): Fix flaky JMS tests
For some reason temporary queues get deleted before the synchronous JMS consumer is able to send a reply message. The error can not be reproduced on local machines but only happens randomly on GitHub CI jobs. Try to overcome this by disabling advisory support and explicitly set the timeout for purging temporary destinations on the JMS broker.