feat(#1215): selective kafka receive action implementation
you can now use selectors when looking for kafka messages.
that significantly improves the value of the feature, because it was previously only possible to receive single messages.

the resolving-mechanism works time-based, because finding messages in a Kafka topic is relatively complicated.
the amount of consumed messages is potentially very high.
therefore, message consumption starts at an offset `Ox = OT-n`.
where `T` is the current timestamp and `n` is the maximum timespan in which the wanted event is expected to have been published.

example usage in java:

```java
@Test
public class KafkaEndpointIT extends TestNGCitrusSpringSupport {

    @Autowired
    private KafkaEndpoint helloKafkaEndpoint;

    @Test
    @CitrusTest
    public void findKafkaEvent_headerEquals_citrus_DSL() {
        var body = "findKafkaEvent_headerEquals_citrus_DSL";

        var key = "Name";
        var value = "Bilbo";

        when(
                send(helloKafkaEndpoint)
                        .message(new KafkaMessage(body).setHeader(key, value))
        );

        then(
                receive(helloKafkaEndpoint)
                        .selector(
                                kafkaMessageFilter()
                                        .eventLookbackWindow(Duration.ofSeconds(1L))
                                        .kafkaMessageSelector(kafkaHeaderEquals(key, value))
                                        .build()
                        )
                        .getMessageBuilderSupport()
                        .body(body)
        );
    }
}
```